learning/requirement.txt
pretty_midi
numpy
librosa
matplotlib
yt-dlp
moviepy
imageio==2.30.0
keras
tensorflow
beautifulsoup4
requests
azure-cli